refactor: move exercise checker to the exercises folder
