+++ src/exercises/checker_test.py
@@ -1,5 +1,5 @@
 import pytest
-from src.exercise_checker.checker import ExerciceFailed, check_exercise
+from src.exercises.checker import ExerciceFailed, check_exercise
 
 
 async def test_protostar_test_checker_success():
